feat: create and user hook with functions user flow
