+++ frontend/src/app/auth/reset-password/page.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 'use client';
 
 import type React from 'react';
@@ -7,79 +6,21 @@ import Link from 'next/link';
 import { Button } from '@/components/ui/button';
 import { Label } from '@/components/ui/label';
 import { BookOpen, ArrowLeft, Mail, LoaderCircleIcon } from 'lucide-react';
-import { toast } from '../../../hooks/use-toast';
-import { supabase } from '../../../lib/supabase';
-import { useRouter } from 'next/navigation';
 import { ToggleEye } from '../../../components/common/ToggleEye';
-import { mapSupabaseError } from '../../../lib/map-errors';
+import { useUserAuth } from '../../../hooks/use-user-auth';
 
 export default function ForgotPassword() {
   const [password, setPassword] = useState('');
-  const [loading, setLoading] = useState(false);
-  const router = useRouter();
+  const { setSession, resetPassword, loading } = useUserAuth();
 
   useEffect(() => {
-    const hash = window.location.hash;
-    const params = new URLSearchParams(hash.replace('#', '?'));
-    const access_token = params.get('access_token');
-    const refresh_token = params.get('refresh_token');
-
-    if (access_token && refresh_token) {
-      supabase.auth
-        .setSession({
-          access_token,
-          refresh_token,
-        })
-        .then(({ error }) => {
-          if (error) {
-            toast({
-              variant: 'success',
-              description: 'Erro ao definir sessão',
-            });
-          }
-
-          setLoading(false);
-        });
-    } else {
-      toast({
-        variant: 'error',
-        description: 'Tokens não encontrados na URL',
-      });
-      setLoading(false);
-    }
+    setSession();
   }, []);
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
 
-    try {
-      setLoading(true);
-
-      const { error } = await supabase.auth.updateUser({ password });
-
-      if (error) {
-        const message = mapSupabaseError(error);
-        toast({
-          variant: 'error',
-          description: message,
-        });
-      } else {
-        toast({
-          variant: 'success',
-          description: 'Senha redefinida com sucesso!',
-        });
-
-        router.push('/auth/signin');
-      }
-    } catch (err: any) {
-      const message = mapSupabaseError(err.error ?? err);
-      toast({
-        variant: 'error',
-        description: message,
-      });
-    } finally {
-      setLoading(false);
-    }
+    resetPassword(password);
   };
 
   return (
